More example cleanup
This example was showing a feature of `getDevices()` that no longer exists. It was also chaining promises incorrectly with a block expecting the result of getDevices firing after a call to setDoorOpen. This cleans all of the example code up to run as-is after replacing email and password (and possibly also the `desiredDeviceName`).