Fix examples and getDevices documentation (#15)
* Update example to work with api v5 changes

* More example cleanup

This example was showing a feature of `getDevices()` that no longer exists. It was also chaining promises incorrectly with a block expecting the result of getDevices firing after a call to setDoorOpen. This cleans all of the example code up to run as-is after replacing email and password (and possibly also the `desiredDeviceName`).

* Update readme

Updated `getDevices()` docs to no longer say that it supports type IDs to filter by.